feat(upload): Add support for uploading files to a nexus repository
This commit adds support for uploading files by introducing a new command-line argument `-r` or `--repository`. The default value is set to "depot-local". Additionally, the code changes include the addition of a new module `upload.rs` in the operations directory.

The `execute_curl_request` function in `network_utils.rs` has been modified to accept an additional argument `args`, which is a HashMap containing upload file information. If the `"upload-file"` key is present in the HashMap, the function enables file upload and reads the contents of the file using a read function.

These changes enable users to upload files as part of their operations.